SW-3893 Add tabs to tracking v2 withdrawals page
- refactor existing withdrawals page to extract out the table content into WithdrawalsTable
- create tabs in v2 plantings and withdrawals
- reuse WithdrawalsTable in withdrawal history tab
- instantiate empty plantings component
